Move SteemKeychainError into errors directory
