+++ src/ngx-steem-keychain.ts
@@ -1,4 +1,4 @@
 export { SteemKeychainModule } from './steem-keychain.module';
-export { SteemKeychainError } from './steem-keychain-error';
+export { SteemKeychainError } from './errors/steem-keychain-error';
 export { Response } from './response';
 export * from './steem-keychain.service';
